+++ src/lib/sanity/schemas/element/casting/elementCastingRegistry.ts
@@ -1,9 +1,9 @@
-import { elementTextSingleLineCastingFields } from "./elementTextSingleLine";
-import { elementImageCastingFields } from "./elementImage";
-import { elementTextBlockCastingFields } from "./elementTextBlock";
-import { elementRichTextCastingFields } from "./elementRichText";
-import { elementButtonCastingFields } from "./elementButton";
-import { elementSVGCastingFields } from "./elementSVG";
+import { elementTextSingleLineCastingFields } from "../elementTextSingleLine";
+import { elementImageCastingFields } from "../elementImage";
+import { elementTextBlockCastingFields } from "../elementTextBlock";
+import { elementRichTextCastingFields } from "../elementRichText";
+import { elementButtonCastingFields } from "../elementButton";
+import { elementSVGCastingFields } from "../elementSVG";
 
 export const elementCastingRegistry: Record<string, unknown[]> = {
   elementTextSingleLine: elementTextSingleLineCastingFields,